# Reduce allocations in `ReducedDensityMatrix`