--- a/src/DictVectors/pdvec.jl
+++ b/src/DictVectors/pdvec.jl
@@ -531,7 +531,20 @@ function Base.mapreduce(f::F, op::O, t::PDVecIterator; kwargs...) where {F,O}
     return merge_remote_reductions(t.vector.communicator, op, result)
 end
 
+"""
+    sum_mutating!(accumulator, [f! = add!], keys(::PDVec); [init])
+    sum_mutating!(accumulator, [f! = add!], values(::PDVec); [init])
+    sum_mutating!(accumulator, [f! = add!], pairs(::PDVec); [init])
 
+Perform a parallel sum on [`PDVec`](@ref)s for vector-valued results while minimizing
+allocations. The result of the sum will be added to `accumulator` and stored in
+`accumulator`. MPI-compatible. If f! is provided, it must accept two arguments, the first
+being the accumulator and the second the element of the iterator. Otherwise, add! is used.
+
+If provided, `init` must be a neutral element for `+` and of the same type as `accumulator`.
+
+See also [`mapreduce`](@ref).
+"""
 function Interfaces.sum_mutating!(accu, f!, iterator::PDVecIterator; kwargs...)
     interim_result = _sum_non_mutating(accu, f!, iterator; kwargs...)
     add!(accu, interim_result)
@@ -543,8 +556,8 @@ function _sum_non_mutating(accu, f!, iterator::PDVecIterator; kwargs...)
     interim_result = Folds.mapreduce(
         +, Iterators.filter(!isempty, iterator.vector.segments); kwargs...
     ) do segment
-        # each segment/thread gets is own copy of the accumulator such that thread can work,
-        # on it independently. Later these are merged
+        # each segment gets is own copy of the accumulator such that each thread can work
+        # on it independently. Later the accumulators are merged.
         sum_mutating!(zero(accu), f!, iterator.selector(segment))
     end
     return merge_remote_reductions(iterator.vector.communicator, +, interim_result)