--- a/src/Hamiltonians/Hamiltonians.jl
+++ b/src/Hamiltonians/Hamiltonians.jl
@@ -99,7 +99,7 @@ if VERSION < v"1.10"
     # used for ReducedDensityMatrix
     function hermitianpart!(A)
         A .= (A + A') / 2
-        return A
+        return Hermitian(A)
     end
 end