## New features

* New function `sum_mutating!` allows parallel reductions from `PDVec`s over vector-valued results while minimizing allocations.

## Changed behavior

Allocations and GC load is reduced leading to a 1.5x speedup when calculating a  dot product with a `ReducedDensityMatrix(2)` on four threads.

**Old code:**
```julia-repl
julia> @benchmark dot($v2, $rdmop, $v2)
BenchmarkTools.Trial: 6 samples with 1 evaluation.
 Range (min … max):  843.080 ms … 953.016 ms  ┊ GC (min … max): 15.06% … 20.26%
 Time  (median):     900.043 ms               ┊ GC (median):    18.58%
 Time  (mean ± σ):   900.078 ms ±  44.495 ms  ┊ GC (mean ± σ):  18.69% ±  2.93%

  █           █     █                          █      █       █  
  █▁▁▁▁▁▁▁▁▁▁▁█▁▁▁▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█▁▁▁▁▁▁█▁▁▁▁▁▁▁█ ▁
  843 ms           Histogram: frequency by time          953 ms <

 Memory estimate: 2.76 GiB, allocs estimate: 77248.

```

**New code (this PR):**
```julia-repl
julia> @benchmark dot($v2, $rdmop, $v2)
BenchmarkTools.Trial: 9 samples with 1 evaluation.
 Range (min … max):  516.414 ms … 615.814 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     593.440 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   587.321 ms ±  30.741 ms  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █                                  █  █  █     █   █      ███  
  █▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█▁▁█▁▁█▁▁▁▁▁█▁▁▁█▁▁▁▁▁▁███ ▁
  516 ms           Histogram: frequency by time          616 ms <

 Memory estimate: 1.10 MiB, allocs estimate: 55.

julia> length(v2)
12870
```